Move ISR callbacks to header again: They're static in order to prevent linker errors when Encoder.h is included from multiple translation units. This does result in slightly higher memory usage if Encoder::begin is called in multiple translation units. In the next commit, i'll probably transform the array of function pointers into a function that returns the right function pointer, so this should fix the extra RAM usage (probably not ROM, though).
AFAIK, this is the only valid way to do it if you want to be able to decide
whether to use interrupts or not by defining a macro before including
the header.
In the cpp file, you don't know if interrupts were enabled or not so you
cannot conditionally include the ISRs there.
I think including the definition of the Encoder class in the cpp file
still technically violates the ODR, but it's probably fine since it
isn't actually used there.